+++ src/02_json_reflection/main.cc
@@ -8,7 +8,7 @@ using magic_io::to_json;
 struct User {
   std::string name;
   std::string status;
-  uint64_t age;
+  int age;
 };
 
 int main() {
